- more time fixes - KALEEE - inversing of switch - erection implemented - translator implemented

+++ termocycler-main/logger/Time.cpp
@@ -12,27 +12,27 @@ private:
 	const static unsigned long hoursInDay = 24UL;
 
 public:
-	int days() {
+	unsigned long days() {
 		unsigned long divider = millisecondsInSecond * secondsInMinute * minutesInHour
 				* hoursInDay;
 		return millis() / divider;
 	}
 
-	int hours() {
+	unsigned long hours() {
 		unsigned long divider = millisecondsInSecond * secondsInMinute * minutesInHour;
 		return (millis() / divider) % hoursInDay;
 	}
 
-	int minutes() {
+	unsigned long minutes() {
 		unsigned long divider = millisecondsInSecond * secondsInMinute;
 		return (millis() / divider) % minutesInHour;
 	}
 
-	int seconds() {
+	unsigned long seconds() {
 		return (millis() / this->millisecondsInSecond) % this->secondsInMinute;
 	}
 
-	int milliseconds() {
+	unsigned long milliseconds() {
 		return millis() % this->millisecondsInSecond;
 	}
 };
